Добавит контроллер
